fix(*): few edge cases handles
